+++ blocklets/ai-runtime/CHANGELOG.md
@@ -1,7 +1,10 @@
+## 0.4.82 (2024-10-18)
+
+- feat: support tracking pageView and events for GA
+
 ## 0.4.81 (2024-10-17)
 
-- chore: add llm adapter global setting
-- feat: add global config for resource blocklet
+- feat: support preview agent and components
 
 ## 0.4.80 (2024-10-17)